refactor(app): extract probeTarget; RunCycle reads as walk→synthesize→reconcile→age (#153)
